+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+v0.3.2 (2024-01-07)
+-------------------
+- edit column view from menu
+- summary column in responsive mode
+- fixed reponsive layout details
+
 v0.3.1 (2024-01-01)
 -------------------
 - global table search